requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.9
distlib==0.3.9
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.51.0
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ipython==8.24.0
ipywidgets==8.1.2
isort==5.13.2
itsdangerous==2.2.0
jcopml==1.2.7
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
luwiji==1.2.3
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numpy==1.26.4
openai==1.54.4
opencv-python==4.10.0.82
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
platformdirs==4.3.6
prompt-toolkit==3.0.43
propcache==0.2.0
psutil==6.1.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyaml==24.4.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pylint==3.3.1
pyparsing==3.1.2
pyperclip==1.8.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.0
scikit-optimize==0.10.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.13.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.3.29
ultralytics-thop==2.0.11
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.10
xlrd==2.0.1
yarl==1.17.1
